Don't delete records in LDAP immediately
.. instead reset all the account flags and mark the record as deleted.
This ensures that in the PH the records doesn't retain the access
rights. After the PH has marked the record as synchronized we delete
records marked synced *and* deleted in LDAP in a garbage collection
step.

Note that the test-driver now must set the etd timestamp for records
marked as deleted. The test-driver now stops if an error occors during
some step in the inner loop.