Add -f option to initial load
This changes the algorithm: If we get a row with a pk_uniqueid that
already exists with a different CN we don't asume the CN is changed but
create an entry in LDAP with duplicate pk_uniqueid and different CN.
This adds test data for the new semantics. It also changes the key in
the test database to *not* be the pk_uniqueid.